feat(service instance): Support for refund and renewal of service instances.
1. Support refund for the latest order.
2. Support refund for all orders
under a service instance.
3. Support renewal of service instances.

+++ boost.serverless/src/main/java/org/example/task/RefundOrderTask.java
@@ -1,17 +1,17 @@
 /*
-*Copyright (c) Alibaba Group;
-*Licensed under the Apache License, Version 2.0 (the "License");
-*you may not use this file except in compliance with the License.
-*You may obtain a copy of the License at
+ *Copyright (c) Alibaba Group;
+ *Licensed under the Apache License, Version 2.0 (the "License");
+ *you may not use this file except in compliance with the License.
+ *You may obtain a copy of the License at
 
-*   http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
 
-*Unless required by applicable law or agreed to in writing, software
-*distributed under the License is distributed on an "AS IS" BASIS,
-*WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*See the License for the specific language governing permissions and
-*limitations under the License.
-*/
+ *Unless required by applicable law or agreed to in writing, software
+ *distributed under the License is distributed on an "AS IS" BASIS,
+ *WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *See the License for the specific language governing permissions and
+ *limitations under the License.
+ */
 
 package org.example.task;
 
@@ -25,6 +25,7 @@ import org.example.common.constant.TradeStatus;
 import org.example.common.dataobject.OrderDO;
 import org.example.common.dto.OrderDTO;
 import org.example.common.helper.OrderOtsHelper;
+import org.example.common.utils.DateUtil;
 
 import java.util.Collections;
 import java.util.concurrent.CountDownLatch;
@@ -32,7 +33,7 @@ import java.util.concurrent.ScheduledExecutorService;
 
 @Builder
 @Slf4j
-public class RefundOrderTask implements Runnable{
+public class RefundOrderTask implements Runnable {
 
     private OrderOtsHelper orderOtsHelper;
 
@@ -63,14 +64,22 @@ public class RefundOrderTask implements Runnable{
                 OrderDO orderDO = new OrderDO();
                 orderDO.setTradeStatus(TradeStatus.REFUNDED);
                 orderDO.setOrderId(orderId);
-                Boolean alipaySuccess = baseAlipayClient.refundOrder(orderId, Double.parseDouble(String.format("%.2f", refundAmount)), refundId);
-                DeleteServiceInstancesRequest deleteServiceInstancesRequest = new DeleteServiceInstancesRequest();
-                deleteServiceInstancesRequest.setServiceInstanceId(Collections.singletonList(serviceInstanceId));
-                DeleteServiceInstancesResponse deleteServiceInstancesResponse = computeNestSupplierClient.deleteServiceInstance(deleteServiceInstancesRequest);
-                Boolean updateOrder = orderOtsHelper.updateOrder(orderDO);
-                log.info("Alipay refund {}. Update order {}. Delete service instance status code = {}", alipaySuccess, updateOrder, deleteServiceInstancesResponse.getStatusCode());
-                log.info("Order refund success. order id = {}.", orderDO.getOrderId());
+                Boolean alipaySuccess = Boolean.TRUE;
+                if (Double.parseDouble(String.format("%.2f", refundAmount)) > 0) {
+                    alipaySuccess = baseAlipayClient.refundOrder(orderId, Double.parseDouble(String.format("%.2f", refundAmount)), refundId);
+                }
+                //todo 删除服务实例+更新计算巢endTime
                 OrderDTO order = orderOtsHelper.getOrder(orderId, null);
+                Long currentLocalDateTimeMillis = DateUtil.getCurrentLocalDateTimeMillis();
+                if (shouldDeleteServiceInstance(currentLocalDateTimeMillis, order)) {
+                    DeleteServiceInstancesRequest deleteServiceInstancesRequest = new DeleteServiceInstancesRequest();
+                    deleteServiceInstancesRequest.setServiceInstanceId(Collections.singletonList(serviceInstanceId));
+                    DeleteServiceInstancesResponse deleteServiceInstancesResponse = computeNestSupplierClient.deleteServiceInstance(deleteServiceInstancesRequest);
+                    log.info("Delete service instance status code = {}", deleteServiceInstancesResponse.getStatusCode());
+                }
+                Boolean updateOrder = orderOtsHelper.updateOrder(orderDO);
+                log.info("Alipay refund {}. Update order {}.", alipaySuccess, updateOrder);
+                log.info("Order refund success. order id = {}.", orderDO.getOrderId());
                 success = order.getTradeStatus() == TradeStatus.REFUNDED;
             } catch (Exception e) {
                 retryCount++;
@@ -87,5 +96,13 @@ public class RefundOrderTask implements Runnable{
         if (!success) {
             log.error("Failed to execute code after multiple retries.");
         }
+    }
+
+
+    private Boolean shouldDeleteServiceInstance(Long currentLocalDateTimeMillis, OrderDTO order) {
+        if (order.getBillingStartDateLong() != null && order.getBillingEndDateLong() != null) {
+            return currentLocalDateTimeMillis >= order.getBillingStartDateLong() && currentLocalDateTimeMillis <= order.getBillingEndDateLong();
+        }
+        return Boolean.FALSE;
     }
 }
